requirements.txt
PyPDF2
openai
python-dotenv>=1.0.0
pydantic
tqdm
tenacity>=8.2.0
litellm>=1.10.0
requests>=2.28.0